chatbot/requirements.txt
# Core
pandas==2.2.2
numpy<2.0.0,>=1.24.0
python-dotenv==1.0.1
streamlit==1.38.0
requests==2.32.3
flask==3.0.0
flask-cors==4.0.0

# LangChain ecosystem
langchain==0.2.16
langchain-community==0.2.12
langchain-core<0.3.0,>=0.2.38
langchain-openai==0.1.22
langchain-groq==0.1.4
langchain-huggingface==0.0.2
langgraph==0.6.7

# Vector databases
chromadb==0.5.5
faiss-cpu==1.8.0.post1

# Embeddings
sentence-transformers==3.1.0
tiktoken==0.7.0
torch==2.4.1
huggingface-hub==0.24.6

# PDF / File handling
pypdf==4.3.1
PyMuPDF==1.24.9

# Google AI
google-generativeai==0.8.0
